requirements.txt
google-adk==1.14.0
langchain-community==0.3.27
wikipedia==1.4.0
google-cloud-datastore==2.19.0
python-dotenv==1.0.1